PythonKDCServer/requirements.txt
firebase-admin
pycryptodome